fix(routes): move settings routes to admin
