+++ server/routes/admin/index.js
@@ -1,8 +1,9 @@
 'use strict';
 
-const actionRoutes = require('./action');
+const actionRoutes = require('./action-routes');
+const settingsRoutes = require('./settings-routes');
 
 module.exports = {
 	type: 'admin',
-	routes: [...actionRoutes],
+	routes: [...actionRoutes, settingsRoutes],
 };
